+++ tests/test_objects.py
@@ -355,3 +355,21 @@ class TestGridFsObjects(unittest.TestCase):
                          "`optest` is the only expected file and we received %s" % listed_files)
 
         yield gfs.delete(_id)
+
+    @defer.inlineCallbacks
+    def test_GridFsIndexesCreation(self):
+        """ Tests gridfs indexes creation"""
+        conn = yield txmongo.MongoConnection(mongo_host, mongo_port)
+        db = conn.test
+        yield self._drop_gridfs(db)
+
+        # Create a new GridFS instance should trigger indexes creation
+        gfs = GridFS(db)
+
+        # Multiple calls should return multiple defer not to mix between them
+        self.assertNotEqual(gfs.get_indexes_created_defer(),
+                            gfs.get_indexes_created_defer())
+
+        yield gfs.get_indexes_created_defer()
+
+        yield conn.disconnect()