+++ docs/source/NEWS.rst
@@ -17,6 +17,9 @@ Features
 
 - ``insert_many()`` is now able to insert more than 1000 documents and more than 16Mb of documents at once.
 - GridFS's default ``chunkSize`` changed to 255kB, to avoid the overhead with usePowerOf2Sizes option.
+- Add ``GridFS.get_indexes_created_defer`` to obtain a defer on the creation of the current
+  GridFS instance's indexes
+- GridFS create indexes for the ``files`` collection in addition to the ``chunks`` one
 
 Release 16.0.1 (2016-03-03)
 ---------------------------